[TASK-321] Extract HTML generator functions into tusk-dashboard-html.py
